+++ packages/oauth2/tests/utils/index.ts
@@ -1,5 +1,4 @@
-import type { JWTPayload, CustomPayload } from "@src/types";
-export { generateHS256JWS, generateRS256JWS } from "@tests/utils/generateJWS";
+import { JWTPayload } from "@src/types";
 
 const defaultPayload: JWTPayload = {
   aud: ["oauth2"],
@@ -50,10 +49,4 @@ vwIDAQAB
 
 const defaultSecret = "c9ab0fdc-b2dc-47ad-933b-87cf1b180ab5";
 
-export {
-  defaultPayload,
-  defaultAsymmetricAlgoKeyPair,
-  defaultSecret,
-  JWTPayload,
-  CustomPayload,
-};
+export { defaultPayload, defaultAsymmetricAlgoKeyPair, defaultSecret };